Reorganize project structure a little. Mainly, docker and compose, migrations manager, go modules, package structure.
