+++ api/rest.go
@@ -1,6 +1,4 @@
-// rest.go
-
-package main
+package api
 
 import (
 	"database/sql"
@@ -9,34 +7,35 @@ import (
 	"strconv"
 
 	"github.com/gorilla/mux"
-	_ "github.com/lib/pq"
+
+	"github.com/Bnei-Baruch/gxydb-api/pkg/httputil"
 )
 
 func (a *App) getGroups(w http.ResponseWriter, r *http.Request) {
 	files, err := getGroups(a.DB)
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
-	respondWithJSON(w, http.StatusOK, files)
+	httputil.RespondWithJSON(w, http.StatusOK, files)
 }
 
 func (a *App) getRooms(w http.ResponseWriter, r *http.Request) {
 	files, err := getRooms(a.DB)
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
-	respondWithJSON(w, http.StatusOK, files)
+	httputil.RespondWithJSON(w, http.StatusOK, files)
 }
 
 func (a *App) getUsers(w http.ResponseWriter, r *http.Request) {
 	files, err := getUsers(a.DB)
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
-	respondWithJSON(w, http.StatusOK, files)
+	httputil.RespondWithJSON(w, http.StatusOK, files)
 }
 
 func (a *App) getRoom(w http.ResponseWriter, r *http.Request) {
@@ -44,7 +43,7 @@ func (a *App) getRoom(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	id, err := strconv.Atoi(vars["id"])
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
 	i.Room = id
@@ -52,13 +51,13 @@ func (a *App) getRoom(w http.ResponseWriter, r *http.Request) {
 	if err := i.getRoom(a.DB); err != nil {
 		switch err {
 		case sql.ErrNoRows:
-			respondWithError(w, http.StatusNotFound, "Not Found")
+			httputil.RespondWithError(w, http.StatusNotFound, "Not Found")
 		default:
-			respondWithError(w, http.StatusInternalServerError, err.Error())
+			httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		}
 		return
 	}
-	respondWithJSON(w, http.StatusOK, i)
+	httputil.RespondWithJSON(w, http.StatusOK, i)
 }
 
 func (a *App) getUser(w http.ResponseWriter, r *http.Request) {
@@ -69,13 +68,13 @@ func (a *App) getUser(w http.ResponseWriter, r *http.Request) {
 	if err := i.getUser(a.DB); err != nil {
 		switch err {
 		case sql.ErrNoRows:
-			respondWithError(w, http.StatusNotFound, "Not Found")
+			httputil.RespondWithError(w, http.StatusNotFound, "Not Found")
 		default:
-			respondWithError(w, http.StatusInternalServerError, err.Error())
+			httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		}
 		return
 	}
-	respondWithJSON(w, http.StatusOK, i)
+	httputil.RespondWithJSON(w, http.StatusOK, i)
 }
 
 func (a *App) postRoom(w http.ResponseWriter, r *http.Request) {
@@ -83,24 +82,24 @@ func (a *App) postRoom(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	id, err := strconv.Atoi(vars["id"])
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
 	i.Room = id
 
 	d := json.NewDecoder(r.Body)
 	if err := d.Decode(&i); err != nil {
-		respondWithError(w, http.StatusBadRequest, "Invalid resquest payload")
+		httputil.RespondWithError(w, http.StatusBadRequest, "Invalid request payload")
 		return
 	}
 
 	defer r.Body.Close()
 
 	if err := i.postRoom(a.DB); err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
-	respondWithJSON(w, http.StatusOK, map[string]string{"result": "success"})
+	httputil.RespondWithJSON(w, http.StatusOK, map[string]string{"result": "success"})
 }
 
 func (a *App) postUser(w http.ResponseWriter, r *http.Request) {
@@ -108,17 +107,17 @@ func (a *App) postUser(w http.ResponseWriter, r *http.Request) {
 
 	d := json.NewDecoder(r.Body)
 	if err := d.Decode(&i); err != nil {
-		respondWithError(w, http.StatusBadRequest, "Invalid resquest payload")
+		httputil.RespondWithError(w, http.StatusBadRequest, "Invalid request payload")
 		return
 	}
 
 	defer r.Body.Close()
 
 	if err := i.postUser(a.DB); err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
-	respondWithJSON(w, http.StatusOK, map[string]string{"result": "success"})
+	httputil.RespondWithJSON(w, http.StatusOK, map[string]string{"result": "success"})
 }
 
 func (a *App) deleteRoom(w http.ResponseWriter, r *http.Request) {
@@ -126,16 +125,16 @@ func (a *App) deleteRoom(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	id, err := strconv.Atoi(vars["id"])
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
 	i.Room = id
 
 	if err := i.deleteRoom(a.DB); err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
-	respondWithJSON(w, http.StatusOK, map[string]string{"result": "success"})
+	httputil.RespondWithJSON(w, http.StatusOK, map[string]string{"result": "success"})
 }
 
 func (a *App) deleteUser(w http.ResponseWriter, r *http.Request) {
@@ -144,8 +143,8 @@ func (a *App) deleteUser(w http.ResponseWriter, r *http.Request) {
 	i.ID = vars["id"]
 
 	if err := i.deleteUser(a.DB); err != nil {
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		httputil.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
-	respondWithJSON(w, http.StatusOK, map[string]string{"result": "success"})
+	httputil.RespondWithJSON(w, http.StatusOK, map[string]string{"result": "success"})
 }
