+++ go.mod
@@ -1,0 +1,15 @@
+module github.com/Bnei-Baruch/gxydb-api
+
+go 1.13
+
+require (
+	github.com/coreos/go-oidc v2.2.1+incompatible
+	github.com/golang-migrate/migrate v3.5.4+incompatible // indirect
+	github.com/gorilla/handlers v1.4.2
+	github.com/gorilla/mux v1.7.4
+	github.com/lib/pq v1.3.0
+	github.com/pquerna/cachecontrol v0.0.0-20180517163645-1555304b9b35 // indirect
+	golang.org/x/crypto v0.0.0-20200403201458-baeed622b8d8 // indirect
+	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d // indirect
+	gopkg.in/square/go-jose.v2 v2.4.1 // indirect
+)
